Change FIND of empty series to match, not fail
Historically searching for empty series in a series would fail:

    rebol2>> find "abc" ""
    == none

This is inconsistent with what it would seem you would want, e.g. in
PARSE when a string is empty:

    >> parse "abc" ["" "abc"]
    == "abc"

It seems that the ability to uniquely express the intent of BLANK! vs.
empty string gives FIND the ability to discern the two, offering a
means of opting out -and- a means of opting in:

    >> find "abc" _
    ; null

    >> find "abc" ""
    == "abc"

This makes the change (tests forthcoming, but it needed to be done to
untangle something else).