Transition UPARSE to Isotope Protocol
When initially writing UPARSE there was no protocol for managing return
states like invisibility, or distinguishing NULL and NULL-2.  So it
used quoted values and `[]` to indicate invisibility, with ' meaning
"heavy null" (successful rule) and null meaning "light null" (match
failure).

This put pressure on developing a systemic answer for reading the
invisible and null-2 state of expressions.  This commit makes the
changes to the system to get the features of @(...) working more with
isotopes, and retrofits UPARSE to use it.

It's a work in progress--as ideas about how to rethink isotopes came
up during development, that should be applied before worrying too much
about getting the tests in order.