VOID! => BAD-WORD! Isotope Concept
This is a systemic overhaul that rethinks the type that was called VOID!
into BAD-WORD!...a type that it is legal to have execute in source,
where it becomes a "mean" version that can't be passed to parameters
that are not @literal.

The changes are too numerous to fully document in this commit message,
so are documented on forum posts (and in the code comments).  It is a
work in progress, as the return convention solidifies.