Switch meanings of VOID->NONE, NIHIL->VOID
This is another massive change which is kind of hard to break apart.
What was previously known as the `<void>` return convention is now
switched over to the `<none>` convention for a BAD-WORD! of ~none~
whose meaning is not conflated with invisibility.  Then it makes
~void~ do more conflation with that invisibility.

So for instance: `return` without an argument now means to return
invisibily.  This won't be allowed if the function return type has
been specified and does not permit invisibility.

This also fixes many tests, though not quite all of them.  The work
is ongoing.