Reclaim ~trash~ term for "Unreadable_Void()"
With the term "void" moving to be used for what were "invisibles",
that means that "unreadable void" would be an "unreadable bad word"
or something otherwise laborious to type.

This moves the term "trash" away from its former use for GC unsafe
cells, and just calls those REB_UNSAFE.  This way, Init_Trash() can
be used to mean initialize a cell with a form of the GC-safe word
~trash~ that happens to assert if read from in the debug build.
(This doesn't happen with all ~trash~ bad words, just those that were
made with the Init_Trash() method.)

It also tidies up some of the potential pitfalls which would have
been caused by doing trash initializations only in debug builds for
performance reasons, by making the idea that there is no such thing
as initialization which might transfer a cell from a GC-unsafe state
to a GC-safe one in debug builds only.  So either the REB_UNSAFE mode
is used with no GC safety, or the trash mode is used and it writes
safe bits in both debug and release builds.