New rules for APPEND/INSERT/CHANGE
This implements the proposed rules that APPEND of a QUOTED! will
remove one level of quoting and append the value as is (no splicing),
and that APPEND of a BLANK! is a no-op, while APPEND of NULL is
an error.

It also implements that evaluated types won't append without either
quoting them (or using the @ literalizing operator) or using /ONLY.
This error is helpful in catching the kind of code that might lose
information on QUOTED!.  Allowing inert data types is a compromise so
that things that aren't building what might be "code" don't have to
pay for the overhead.