Change @word family to ^word family
Initially the @ types were introduced to be inert, used for dialects
but also likely used for datatypes.

Over time, they came to have purpose in signaling the intent to use
a value literally... e.g. to evaluate to a quoted version of the value
passed.

As this purpose became critical and cemented, the rather hefty look
of the @ seemed to be a mismatch for the function.  Code that wanted
to convey a notion something like quoting was carrying a giant block
of a symbol.

In light of this, the caret character seemed like a better fit.  It's as
if it were pointing to the place where it would be appending a quote
after evaluation and saying "I'll be putting a quote here..."