JUST => THE
Historical Rebol called the operation that took its parameter
unevaluated and returned it QUOTE:

    rebol2>> quote x
    == x

    rebol2>> quote (10 + 20)
    == (10 + 20)

This name was needed in Ren-C for a function that would add a generic
quoting level to its argument.  There was no better name than QUOTE.

So a search was on to find a new name for this unevaluating operator.
LIT and LITERAL and LITERALLY were tried, and then the term JUST.

Motivations for a different use for JUST came along, which put the
search back on.  The short and clear word THE seems an ideal choice
for this particular simple operation:

    >> the x
    == x

    >> the (10 + 20)
    == (10 + 20)

This makes that change and reclaims JUST for a different noble cause.