Early boot FAIL
FAIL is a somewhat complex function that doesn't need to be terribly
fast.  Giving good failure messages and behavior was thus delegated
to be the responsbility of usermode code, that could grow and improve
as time went on.

However, it can be difficult to debug early boot failures when it is
not provided.  A temporary hack was accidentally committed that made
it easier to get at the reason for the failure.  This deletes that
hack and adds a simple native to take over the function.