DEFAULT only overwrites "unfriendly" unset
The concept behind BAD-WORD!s is to have two modes when stored in
variables but only one when stored in blocks...to be either "friendly"
or "unfriendly.

Having a friendly mode indicates that you can use the variable for
code structuring purposes.  A variable holding a friendly ~unset~
is thus not really "unset", and so DEFAULT should not overwrite it.

Additionally, BAD-WORD!s can be used to convey all sorts of meanings
besides unsetness.  Those meanings may be interesting and DEFAULT
should not overwrite them.

(This also makes DEFAULT overwrite ~void~, which may or may not be
a good idea.  Review.)

Also reorganizes the "curse words" somewhat better, with the "trash"
intent broken out into a separate file.