Use ~null~ for the "NULL-2" isotope
"null isotopes" were conceived prior to BAD-WORD! isotopes.  So the
idea of having the BAD-WORD! of `~null~` in its evaluated form act
as the decaying isotope wasn't thought of.

But now BAD-WORD! isotopes exist and have some instance-specific
behavior (e.g. interconvertibility of ~void~ and the absence of
values).  The idea of interconverting ~null~ in its "unstable" form
thus has precedent.

This leverages the idea to make ~null~ a more visceral form of the
null isotope.  It decays several places, like assignment and when
it is fulfilled as a "normal" parameter.

Since ~null~ is still a somewhat irregular datatype, it's less of an
issue to allow `nullptr` passed in via the API to become a ~null~.
This makes it easier to do things like test refinements for truthiness
without needing to use rebQ or rebDidQ.  This paves the way for
the already-planned elimination of the rebXxxQ() set of APIs.