Change rebSpell() and rebBytes() to blank opt-out
The APIs for rebSpell() and rebBytes() tried to be helpful by passing
through a NULL input and giving NULL as an output.

But the system philosophy is now generally to consider most NULL inputs
to be errors.  So "blank in, null out" is more common than passing
through null without comment.

This changes to use BLANK!.