Clean up isotope handling, ensure none in blocks
The "isotope" flag was initially conceived as thinking of the BAD-WORD!
that lived in blocks as the "isotope", because it was "heavy" like
the heavy null.  It was used on two types, NULL and BAD-WORD!

With the concept that isotopic null is actually a BAD-WORD! that is
carrying the "unfriendly" form of the bit, this turns around to where
it makes the most sense to just say that isotopes are the unfriendly
forms.  This makes that reversal--although ultimately the bit will
probably be unified with the evaluative bit.

Additionally, there was some leakage of "unfriendly" bits into values
in blocks by means of quoted types.  This makes sure all quoted values
do not have the isotope flag as an invariant, to keep the unquoting
process from putting isotopes in blocks.

It also sorts out the issue of how rebQ() and rebU() interoperate with
isotopes, by creating isotope-aware versions of Quotify and Unquotify.