Make TAG! subdispatching combinator for UPARSE
Historical PARSE matched tags literally, while UPARSE pushes to the idea
that they are better leveraged as "special nouns" to avoid interfering
with the user variables in wordspace.

There is an overall TAG! combinator which looks in the combinator map
for specific tags.  You can replace individual tag combinators or
change the behavior of tags overall completely.

This means you use `to <end>` instead of `to end`.  This can be
overridden, and for the moment aliases of HERE and END are made.

The previous SKIP is tried out under the name `<any>`