Fix POSIX Startup to comply with new FIND rules
The new rule is that FIND (APPEND, SELECT, etc.) of any evaluative
types in the "/ONLY" sense must be done via either quoting it or
turning it into a block.

It is being tested as a rule only for evaluative types at first.

This fixes some overlooked cases in startup of using FIND with WORD!
in a BLOCK! of words.