Fix GOB! comparison
The comparison model in R3-Alpha required serious review:

https://forum.rebol.info/t/comparison-semantics/1318

A simplifying change to comparisons was made to establish a single
CT_XXX() function that returned -1, 0, or 1...with a single "strict"
flag that could be heeded or not.

The GOB! data type had no tests of comparison so it wasn't noticed
that it hadn't been changed.