Accessing known GOB! fields return NULL vs. error
There is a difference in PATH!-based access of fields between when
a field that is "known" but set to null, vs. a field that is not
part of an object at all.  The former returns null, the latter raises
an error at the time of the path access.

R3-Alpha GOB!s historically had allowed accessing fields like "parent"
of a gob with no parent, and returned none in that case.

However, Ren-C was raising an error when it should have returned NULL

This changes the Get_GOB_Var() function to distinguish between asking
for a field that's not part of GOB!s, and asking for a field that is
known but effectively null.