Fix read-only pane usage in GOB!'s TAKE
Ren-C converted the optimized pane "pointer list" implementation into
a plain array of GOB!s for simplicity and interoperability with the
existing GC methods.

As part of that simplification, GOB!'s TAKE would simply call the
plain TAKE operation on the array held by the GOB.  However, it did
not quote the pane.  As a consequence, the array appeared to be
read-only.

This adds a rebQ() to the pane.  It also adds some more comments to
the gob test file.