NULL KEEP fix in SOURCE
The new rules for functions like APPEND (and hence KEEP, which is built
on APPEND) is that NULL is not allowed, and that plain BLANK! is used
to "opt out".  Hence a QUOTED! blank must be used in order to actually
add a blank.

This addresses a longstanding question about the role of NULL as a
protection against mistakes vs. a way to elegantly opt out:

https://forum.rebol.info/t/635/9