Add missing HERE in %sys-load.r parse
The goal is that SET-WORD! be used instead of SET VARNAME in native
PARSE, in the near term (before a unifying merge with UPARSE).

This means previous usages of SET-WORD! to capture the parse position
have to go through a transitional time where they are changed to
SET-WORD! followed by "HERE".  A case in %sys-load.r that was missed
was noticed while fixing another issue.