Add --no-encap option to skip encap checking
The encap feature is currently operational:

https://forum.rebol.info/t/1576

But as written, it can as much as double the startup time.  It could
certainly be faster, but working on that is not currently a priority.

So if startup speed is important, the current suggestion is to use
this new `--no-encap` command line switch.