Patch Redo interaction with RETURN
The handling of "RETURN parameters" needs significant review.  But
they are classified as refinements at the moment.  This means that
the Redo mechanics were thinking that if a return local had been
filled in (by a returning function), then they'd have to be pushed
to the stack to be processed.

Redo mechanics are yet another thing that needs rethinking.  :-/
However, this patches the problem by just not pushing returns.