Make ODBC dialect work with ^xxx evaluativeness
Several changes had affected the ODBC code, in particular the change
to make @xxx values into ^xxx values that were evaluative.  That
broke the expectations of the small "ODBC dialect".

While it is anticipated that @xxx values will return, this just makes
the ^xxx values processed prior to evaluation instead of after a
REDUCE step.