Some ODBC Modernizations
This changes some instances of rebQ() to use the more readable @, and
switches from slashes to dots for field access.

It also takes fields that are conceptually empty but want to be
testable without an error to be NULL more frequently, as BLANK! is
more lenient than desired...opting out of APPEND and FIND and other
operations vs. causing an error.