Encap fix incorrect zip size on Windows
There were a few things that had atrophied in the encap, but one aspect
seems to have been broken from before (but zip just wasn't noticing it).

Stored payloads were being fetched out of the embedding with the wrong
size on Windows, because the rounded-up section size (physical size)
was being used instead of the virtual size that was actually how big
the embedded data was.