Change REPEAT to act like LOOP, FOR => CFOR
The word "repeat" seems to be a good fit for something that isn't
FOR-like, e.g. it does the same thing (e.g. no additional parameter of
which iteration it is on).

LOOP is more generic, and by making the term available it could be
used as the replacement for arity-2 WHILE.  This would mean that WHILE
could be arity-1 and hence consistent in parse and regular code.

This is a first step toward that change by helping migrate all the
LOOP instances to become REPEAT.

Also, the FOR function is changing.  This gives a little hint of what
that is like while still being able to redirect callers with the
wrong convention to the CFOR function.