Rename SYM-XXX to META-XXX
Originally, the @word, @tu.p.le, @pa/th, @(gr o up) and @[bl o ck]
types were inert.  In the puzzling search for a name these inert
entities were called SYM-XXX! types, since they were "symbolic" and
had no evaluator behavior.

An emergent usage of these types was to create so-called "modal
parameters" which allowed functions like APPEND to notice when they
were used literally at a callsite, and have that trigger an option
like /ONLY.

This idea gave way to making it so that instead of being inert, the
types would evaluate and add a quoting level to their product.  Then,
it was the quoting that was noticed by functions like APPEND.

Once that happened, it began to seem that ^word and ^tu.p.le etc.
would be a more appropriate symbol for what was going on.  It not
only added quoting levels, but it would also transition an "isotope"
of a BAD-WORD! to b eing a regular BAD-WORD!.

This concept ultimately seems best named META-XXX!, which goes well
with the usual use of ^ for escaping in programming notations.

There will still likely be the @ forms brought back in the future,
that will be inert.  And their naming is still up in the air.  But
the META name seems a good fit for the ^ forms.