Eliminate integer ranges in PARSE, force REPEAT use
Historically, literal integers in PARSE rules are used to mean the
rule following is repeated N times:

    r3-alpha>> parse "aaa" [3 "a"]
    == #[true]

This shorthand has an extended property, which is to allow ranges of
accepted repetitions.  For instance "between 2 and 3 instances of a"

    r3-alpha>> parse "aaa" [2 3 "a"]
    == #[true]

    r3-alpha>> parse "aa" [2 3 "a"]
    == #[true]

    r3-alpha>> parse "a" [2 3 "a"]
    == #[false]

However, this is problematic when used in the regularized format of a
combinator-based parse.  Because it means these meanings are distinct:

    parse ... [2 3 "a"]
    parse ... [2 [3 "a"]]

This doesn't fit in with UPARSE.  So the feature is being killed in
native parse as preparation for their unification.  If a combinator for
repeat ranges is needed it will have to be its own combinator.

As a workaround, two rules can be used:

    parse ... [2 "a" opt "a"]  ; between 2 and 3 "a"
    parse ... [2 "a" 3 opt "a"]  ; between 2 and 5 "a"

Additionally, this commit makes it so that you must use the REPEAT
rule when invoking a repeat with a repetition.  So instead of:

    >> var: 3
    >> parse "aaa" [var "a"]

You say:

    >> var: 3
    >> parse "aaa" [repeat (var) "a"]
    == #[true]

This resolves some issues of readability in parse code, where a
variable doesn't represent a rule but rather something that indirectly
takes an argument and acts as a rule.